linenoise: Readline style completion
Instead of looping through a completion everytime the tab key is hit,
just display all completions.

If there is a common prefix of completions, add that to the line
and don't display all the completions unless the user hits tab again.

Wordwise completion instead of linewise, word delim is hardcoded to space.